Add PeriodicScheduler tests, extract from default
* Add tests to idependently test PeriodicScheduler; coverage at 100%
* Extract logic specific to PeriodicScheduler from the DefaultScheduler
  class.
